RE 7213_Sort once in splatterplot
Added instance variables to hold the sorted list of boxes
and the workspace name.  The list sorted list will now only
be made if it was not previously made, or if the MD workspace
name changes.
NOTE: This does not yet solve the problem, since the whole
splatter plot factory is rebuilt everytime the user changes
the number of points or percent.  One possible solution would
be for the higher level class to only construct a new splatter
plot factory in case the workspace changes, and just call the
doCreate() method on that factory when the user changes the
number of points or percent.

refs #7213